[IMP] payment: rename of acquirer on provider
Changing the name of model payment.acquirer to payment.provider
and everything that it touches. It is technically incorrect to
use the term "acquirer" for systems that only provide a service
of payment.
After this commit the model payment.acquirer and everything
related to it will be renamed to payment.provider.

Task(on odoo-dev) - 2842088